isolate draw schematic grid
