+++ lib/sch/convert-circuit-json-to-schematic-svg.ts
@@ -1,5 +1,6 @@
 import type { AnyCircuitElement } from "circuit-json"
 import { colorMap } from "lib/utils/colors"
+import { drawSchematicGrid } from "./draw-schematic-grid"
 import { stringify } from "svgson"
 import { createSchematicComponent } from "./svg-object-fns/create-svg-objects-from-sch-component"
 import { createSvgObjectsFromSchDebugObject } from "./svg-object-fns/create-svg-objects-from-sch-debug-object"
@@ -55,78 +56,7 @@ export function convertCircuitJsonToSchematicSvg(
   // Add grid if enabled
   if (options?.grid) {
     const gridConfig = typeof options.grid === "object" ? options.grid : {}
-    const cellSize = gridConfig.cellSize ?? 1
-    const labelCells = gridConfig.labelCells ?? false
-    const gridLines: any[] = []
-
-    // Vertical lines
-    for (let x = Math.ceil(minX); x <= Math.floor(maxX); x += cellSize) {
-      gridLines.push({
-        name: "line",
-        type: "element",
-        attributes: {
-          x1: x.toString(),
-          y1: minY.toString(),
-          x2: x.toString(),
-          y2: maxY.toString(),
-          stroke: colorMap.schematic.grid,
-          "stroke-width": "0.01",
-          "stroke-opacity": "0.5",
-        },
-      })
-    }
-
-    // Horizontal lines
-    for (let y = Math.ceil(minY); y <= Math.floor(maxY); y += cellSize) {
-      gridLines.push({
-        name: "line",
-        type: "element",
-        attributes: {
-          x1: minX.toString(),
-          y1: y.toString(),
-          x2: maxX.toString(),
-          y2: y.toString(),
-          stroke: colorMap.schematic.grid,
-          "stroke-width": "0.01",
-          "stroke-opacity": "0.5",
-        },
-      })
-    }
-
-    // Add cell labels if enabled
-    if (labelCells) {
-      for (let x = Math.ceil(minX); x <= Math.floor(maxX); x += cellSize) {
-        for (let y = Math.ceil(minY); y <= Math.floor(maxY); y += cellSize) {
-          gridLines.push({
-            name: "text",
-            type: "element",
-            attributes: {
-              x: x.toString(),
-              y: y.toString(),
-              fill: colorMap.schematic.grid,
-              "font-size": (cellSize / 6).toString(),
-              "fill-opacity": "0.5",
-            },
-            children: [
-              {
-                type: "text",
-                value: `${x},${y}`,
-                name: "",
-                attributes: {},
-                children: [],
-              },
-            ],
-          })
-        }
-      }
-    }
-
-    svgChildren.push({
-      name: "g",
-      type: "element",
-      attributes: { class: "grid" },
-      children: gridLines,
-    })
+    svgChildren.push(drawSchematicGrid(minX, minY, maxX, maxY, gridConfig))
   }
 
   // Add labeled points if provided
